requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-redis==5.4.0
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.0
typing_extensions==4.12.2
